deprecate function: `selectionFeaturesLayer`
